inflate loading spinner transition time to verify thread blocking
it looks like the wait time for a newly loaded layer to render on the map is getting artificially inflated by the transition time related to changing the opacity of the spinner element.  It makes it clear that something's loading, but there's still a gap after the element disappears before the layer renders, which feels icky.  Current hypothesis: the transition is part of the style-property-changing operation and must finish before `element.style.opacity = '0'` can return; plus that the event-handling occurs in the same thread in which the layers are rendered, so that the delay in the return/evaluation of `element.style.opacity = whatever` is also a delay in the return of the event-listener function and therefore also a delay before the rendering of the newly loaded layer can start.